rename: productservice 패키지 이름을 service로 변경
