+++ src/main/java/com/sparta/easydelivery/product/controller/ProductController.java
@@ -3,13 +3,11 @@ package com.sparta.easydelivery.product.controller;
 import com.sparta.easydelivery.product.dto.ProductRequestDto;
 import com.sparta.easydelivery.product.dto.ProductResponseDto;
 import com.sparta.easydelivery.product.dto.ProductUpdateRequestDto;
-import com.sparta.easydelivery.product.productservice.ProductService;
-import com.sparta.easydelivery.user.entity.User;
+import com.sparta.easydelivery.product.service.ProductService;
 import com.sparta.easydelivery.user.implement.UserDetailsImpl;
 import jakarta.validation.Valid;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
-import org.springframework.boot.autoconfigure.kafka.KafkaProperties.Admin;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
@@ -18,7 +16,6 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
